Reestructuración y organización de 'resources'
