+++ app/mod_profiles/resources/lists/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+
+from .genderList import GenderList
+from .profileList import ProfileList
+from .measurementList import MeasurementList
+from .measurementSourceList import MeasurementSourceList
+from .measurementTypeList import MeasurementTypeList
+from .measurementUnitList import MeasurementUnitList
+
+from .profileLatestMeasurementList import ProfileLatestMeasurementList
+from .profileMeasurementList import ProfileMeasurementList
