+++ app/mod_profiles/resources/lists/genderList.py
@@ -4,17 +4,17 @@ from flask_restful import Resource, reqparse, marshal_with
 from flask_restful_swagger import swagger
 from app.mod_shared.models import db
 from app.mod_profiles.models import *
-from .measurementSourceFields import MeasurementSourceFields
+from app.mod_profiles.resources.fields.genderFields import GenderFields
 
 parser = reqparse.RequestParser()
 parser.add_argument('name', type=str, required=True)
 parser.add_argument('description', type=str)
 
-class MeasurementSourceList(Resource):
+class GenderList(Resource):
     @swagger.operation(
-        notes=u'Retorna todas las instancias existentes de fuente de medición.'.encode('utf-8'),
-        responseClass='MeasurementSourceFields',
-        nickname='measurementSourceList_get',
+        notes=u'Retorna todas las instancias existentes de género.'.encode('utf-8'),
+        responseClass='GenderFields',
+        nickname='genderList_get',
         responseMessages=[
             {
               "code": 200,
@@ -22,26 +22,26 @@ class MeasurementSourceList(Resource):
             }
           ]
         )
-    @marshal_with(MeasurementSourceFields.resource_fields, envelope='resource')
+    @marshal_with(GenderFields.resource_fields, envelope='resource')
     def get(self):
-        measurement_sources = MeasurementSource.query.all()
-        return measurement_sources
+        genders = Gender.query.all()
+        return genders
 
     @swagger.operation(
-        notes=u'Crea una nueva instancia de fuente de medición, y la retorna.'.encode('utf-8'),
-        responseClass='MeasurementSourceFields',
-        nickname='measurementSourceList_post',
+        notes=u'Crea una nueva instancia de género, y la retorna.'.encode('utf-8'),
+        responseClass='GenderFields',
+        nickname='genderList_post',
         parameters=[
             {
               "name": "name",
-              "description": u'Nombre de la fuente de medición.'.encode('utf-8'),
+              "description": u'Nombre del género.'.encode('utf-8'),
               "required": True,
               "dataType": "string",
               "paramType": "body"
             },
             {
               "name": "description",
-              "description": u'Descripción de la fuente de medición.'.encode('utf-8'),
+              "description": u'Descripción del género.'.encode('utf-8'),
               "required": False,
               "dataType": "string",
               "paramType": "body"
@@ -54,11 +54,11 @@ class MeasurementSourceList(Resource):
             }
           ]
         )
-    @marshal_with(MeasurementSourceFields.resource_fields, envelope='resource')
+    @marshal_with(GenderFields.resource_fields, envelope='resource')
     def post(self):
         args = parser.parse_args()
-        new_measurement_source = MeasurementSource(args['name'],
-                                                   args['description'])
-        db.session.add(new_measurement_source)
+        new_gender = Gender(args['name'],
+                            args['description'])
+        db.session.add(new_gender)
         db.session.commit()
-        return new_measurement_source, 201
+        return new_gender, 201
