+++ app/mod_profiles/resources/lists/measurementTypeList.py
@@ -4,17 +4,17 @@ from flask_restful import Resource, reqparse, marshal_with
 from flask_restful_swagger import swagger
 from app.mod_shared.models import db
 from app.mod_profiles.models import *
-from .genderFields import GenderFields
+from app.mod_profiles.resources.fields.measurementTypeFields import MeasurementTypeFields
 
 parser = reqparse.RequestParser()
 parser.add_argument('name', type=str, required=True)
 parser.add_argument('description', type=str)
 
-class GenderList(Resource):
+class MeasurementTypeList(Resource):
     @swagger.operation(
-        notes=u'Retorna todas las instancias existentes de género.'.encode('utf-8'),
-        responseClass='GenderFields',
-        nickname='genderList_get',
+        notes=u'Retorna todas las instancias existentes de tipo de medición.'.encode('utf-8'),
+        responseClass='MeasurementTypeFields',
+        nickname='measurementTypeList_get',
         responseMessages=[
             {
               "code": 200,
@@ -22,26 +22,26 @@ class GenderList(Resource):
             }
           ]
         )
-    @marshal_with(GenderFields.resource_fields, envelope='resource')
+    @marshal_with(MeasurementTypeFields.resource_fields, envelope='resource')
     def get(self):
-        genders = Gender.query.all()
-        return genders
+        measurement_types = MeasurementType.query.all()
+        return measurement_types
 
     @swagger.operation(
-        notes=u'Crea una nueva instancia de género, y la retorna.'.encode('utf-8'),
-        responseClass='GenderFields',
-        nickname='genderList_post',
+        notes=u'Crea una nueva instancia de tipo de medición, y la retorna.'.encode('utf-8'),
+        responseClass='MeasurementTypeFields',
+        nickname='measurementTypeList_post',
         parameters=[
             {
               "name": "name",
-              "description": u'Nombre del género.'.encode('utf-8'),
+              "description": u'Nombre del tipo de medición.'.encode('utf-8'),
               "required": True,
               "dataType": "string",
               "paramType": "body"
             },
             {
               "name": "description",
-              "description": u'Descripción del género.'.encode('utf-8'),
+              "description": u'Descripción del tipo de medición.'.encode('utf-8'),
               "required": False,
               "dataType": "string",
               "paramType": "body"
@@ -54,11 +54,11 @@ class GenderList(Resource):
             }
           ]
         )
-    @marshal_with(GenderFields.resource_fields, envelope='resource')
+    @marshal_with(MeasurementTypeFields.resource_fields, envelope='resource')
     def post(self):
         args = parser.parse_args()
-        new_gender = Gender(args['name'],
-                            args['description'])
-        db.session.add(new_gender)
+        new_measurement_type = MeasurementType(args['name'],
+                                               args['description'])
+        db.session.add(new_measurement_type)
         db.session.commit()
-        return new_gender, 201
+        return new_measurement_type, 201
