+++ app/mod_profiles/resources/lists/profileLatestMeasurementList.py
@@ -4,7 +4,7 @@ from flask_restful import Resource, marshal_with
 from flask_restful_swagger import swagger
 from app.mod_shared.models import db
 from app.mod_profiles.models import *
-from .measurementFields import MeasurementFields
+from app.mod_profiles.resources.fields.measurementFields import MeasurementFields
 
 class ProfileLatestMeasurementList(Resource):
     # Crea una copia de los campos del recurso 'MeasurementView'.
