+++ app/mod_profiles/resources/views/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+from .genderView import GenderView
+from .profileView import ProfileView
+from .measurementView import MeasurementView
+from .measurementSourceView import MeasurementSourceView
+from .measurementTypeView import MeasurementTypeView
+from .measurementUnitView import MeasurementUnitView
