+++ app/mod_profiles/resources/views/measurementView.py
@@ -4,7 +4,7 @@ from flask_restful import Resource, reqparse, marshal_with
 from flask_restful_swagger import swagger
 from app.mod_shared.models import db
 from app.mod_profiles.models import *
-from .measurementFields import MeasurementFields
+from app.mod_profiles.resources.fields.measurementFields import MeasurementFields
 
 parser = reqparse.RequestParser()
 parser.add_argument('datetime', required=True)
